requirements.txt
fastapi~=0.104.1
uvicorn
redis~=5.0.1
python-dotenv~=1.0.0
starlette==0.32.0.post1